polarssl: don't ignore failed ssl_init(), but ASSERT() on failure
Since key_state_ssl_init() returns void, it *must* succeed. By changing the
'if(a) { success } else ignore' into 'ASSERT(a); success', we make sure we
don't fail (which we in practice don't do, because otherwise we would have
had complaining users with hard to trace errors).

This patch just changes the if() to an ASSERT(), but that causes a lot of
indenting changes, making this patch look a lot bigger than it really is.

Signed-off-by: Steffan Karger <steffan@karger.me>